refactor: 멤버 도메인 패키지 변경 및 엔티티 생성을 위한 VO 추가
