+++ module-domain/src/main/kotlin/site/yourevents/member/domain/MemberVO.kt
@@ -1,0 +1,13 @@
+package site.yourevents.member.domain
+
+import site.yourevents.auth.vo.KakaoProfile
+
+data class MemberVO(
+    val nickname: String,
+    val email: String,
+) {
+    companion object {
+        fun from(kakaoProfile: KakaoProfile): MemberVO =
+            MemberVO(kakaoProfile.nickname, kakaoProfile.email)
+    }
+}
